requirements.txt
streamlit==1.44.1
transformers==4.51.3
huggingface-hub==0.30.2
sentence-transformers==3.4.1
faiss-cpu==1.10.0
pypdf==5.4.0
python-dotenv==1.1.0
numpy==2.2.5
dotenv
